Refactored the Namespace resolution mechanism to add thread-context based prefix mappings checked at object construction time. This applies to Namespace-aware objects constructed directly or indirectly within the context of a Namespace-aware container, and accounts for the major use case.
The Expression AST traversal "copy" mechanism, which
resolves unresolved namespace prefixes at a post-construction
point in time, has been made more efficient to ensure that a
copy of an Expression is only used when the copy is not
identical to the original instance.

The Namespace mechanism has been moved to its own
common package spiralcraft.common.Namespace. An empty string
is used to represent the "default" prefix mapping, instead of
requiring an extra method.